Stop using access time for file comparison.
The comparison of access times was introduced with commit 1e39ef331b2b78baec84fdb577d497511cc46bd5.
This leads to the fact that the lld sets the variable isUnderSysroot to false by mistake with strongly
parallelized compiles. (Since the access time of the sysroot folder can be updated between two stats).

This prevents the ScriptParser from correctly resolving the path from a linker script if sysroot is set.